move sparse code to its own module
